schema and examples are now in a separate file than the test file
