+++ tests/examples/holiday.js
@@ -1,0 +1,29 @@
+const schema = {
+  type: 'array',
+  items: [
+    {
+      type: 'string'
+    }
+  ]
+}
+const example = [
+  'Christmas',
+  'Thanksgiving',
+  'Summer',
+  'Fall',
+  'New Year',
+  'Game Day',
+  'Winter',
+  'Spring',
+  'Halloween',
+  'Hanukkah',
+  'Passover',
+  'Easter',
+  'Chinese New Year',
+  '4th of July'
+]
+
+module.exports = {
+  schema,
+  example
+}
