+++ tests/gallon.test.js
@@ -3,6 +3,9 @@
 const { matchers } = require('jest-json-schema')
 expect.extend(matchers)
 
+const schema = require('./examples/gallon').schema
+const example = require('./examples/gallon').example
+
 const { gallon } = require('../filesObjects')
 
 describe('this test prevents to any issues and problems, also to break the structure of gallon data', () => {
@@ -10,25 +13,6 @@ describe('this test prevents to any issues and problems, also to break the struc
     expect(gallon).not.toBe('')
   })
 })
-
-const schema = {
-  properties: {
-    name: { type: 'string' },
-    americanStandart: { type: 'string' },
-    americanStandartOunces: { type: 'string' },
-    volume: { type: 'string' },
-    weight: { type: 'string' }
-  },
-  required: ['name', 'americanStandart', 'americanStandartOunces', 'volume', 'weight']
-}
-
-const example = {
-  name: 'gallon',
-  americanStandart: '16 cups, 4 quarts',
-  americanStandartOunces: '128 ounces',
-  volume: '3.79 liters',
-  weight: 'null'
-}
 
 describe('test gallon json schema', () => {
   it('validates my json', () => {
