+++ tests/grocery.test.js
@@ -6,46 +6,14 @@ const grocery = require(groceryFilePath)
 const { matchers } = require('jest-json-schema')
 expect.extend(matchers)
 
+const schema = require('./examples/grocery').schema
+const example = require('./examples/grocery').example
+
 describe('grocery data files returns array', () => {
   it('these tests prevent any issues and problems, also to break the structure of grocery', () => {
     expect(grocery).not.toBe('')
   })
 })
-
-const schema = {
-  properties: {
-    departments: {
-      type: 'array'
-    },
-    id: {
-      type: 'number'
-    },
-    name: {
-      type: 'string'
-    },
-    img: {
-      type: 'boolean'
-    },
-    desc: {
-      type: 'boolean'
-    },
-    slug: {
-      type: 'boolean'
-    }
-  },
-  required: ['departments', 'id', 'name', 'img', 'desc', 'slug']
-}
-const example =
-{
-  'departments': [
-    'Other'
-  ],
-  'id': 14,
-  'name': 'number-four9',
-  'img': false,
-  'desc': false,
-  'slug': false
-}
 
 describe('tests grocery json schema', () => {
   it('validates my json', () => {
