+++ tests/holiday.test.js
@@ -4,31 +4,14 @@ const { holiday } = require('../files')
 const { matchers } = require('jest-json-schema')
 expect.extend(matchers)
 
+const schema = require('./examples/holiday').schema
+const example = require('./examples/holiday').example
+
 describe('holidays data files returns array', () => {
   it('these tests prevent any issues and problems, also to break the structure of holiday', () => {
     expect(holiday).not.toBe('')
   })
 })
-
-const schema = {
-  type: 'array', items: [{ type: 'string' }]
-}
-const example = [
-  'Christmas',
-  'Thanksgiving',
-  'Summer',
-  'Fall',
-  'New Year',
-  'Game Day',
-  'Winter',
-  'Spring',
-  'Halloween',
-  'Hanukkah',
-  'Passover',
-  'Easter',
-  'Chinese New Year',
-  '4th of July'
-]
 
 describe('tests holiday json schema', () => {
   it('validates my json', () => {
