+++ tests/measurementSystems.test.js
@@ -6,6 +6,13 @@ expect.extend(matchers)
 const { measurementSystemFilePath, generatedMeasurementSystem } = require('../files')
 const measurementSystems = require(measurementSystemFilePath)
 
+const {
+  measurementSystemsSchema,
+  measurementSystemsExample,
+  generatedMeasSystemSchema,
+  generatedMeasSystemExample
+} = require('./examples/measurementSystems')
+
 describe('users data files returns array', () => {
   it('these tests prevent any issues and problems, also to break the structure of measurementSystems', () => {
     expect(measurementSystems).not.toBe('')
@@ -18,39 +25,15 @@ describe('users data files returns array', () => {
   })
 })
 
-let schema = {
-  'properties': {
-    'alias': { 'type': 'string' }
-  },
-  'required': [ 'alias' ]
-}
-
-let example = { 'alias': 'universal' }
 
 describe('MeasurementSystem json schema testing', () => {
   it('validates MeasurementSystem json schema', () => {
-    expect(example).toMatchSchema(schema)
+    expect(measurementSystemsExample).toMatchSchema(measurementSystemsSchema)
   })
 })
 
-schema = {
-  'properties': {
-    'id': { 'type': 'string' },
-    'alias': { 'type': 'string' },
-    'title': { 'type': 'string' }
-  },
-  'required': [ 'id', 'alias', 'title' ]
-
-}
-
-example = {
-  'id': 'fbe4fd00-7a3e-11e9-8ec5-0d5ec94f7bcf',
-  'alias': 'universal',
-  'title': 'Universal'
-}
-
 describe('generatedMeasurementSystem json schema testing', () => {
   it('validates generatedMeasurementSystem json schema', () => {
-    expect(example).toMatchSchema(schema)
+    expect(generatedMeasSystemExample).toMatchSchema(generatedMeasSystemSchema)
   })
 })
