+++ tests/nutritionalAttributes.test.js
@@ -1,14 +1,13 @@
 /* global describe, it, expect */
 'use strict'
-const {
-  matchers
-} = require('jest-json-schema')
+const { matchers } = require('jest-json-schema')
 expect.extend(matchers)
 
-const {
-  nutritions2FilePath
-} = require('../files')
+const { nutritions2FilePath } = require('../files')
 const nutritionalAttributes = require(nutritions2FilePath)
+
+const schema = require('./examples/nutritionalAttributes').schema
+const example = require('./examples/nutritionalAttributes').example
 
 describe('nutritionalAttributes data files returns array', () => {
   it('these tests prevent any issues and problems, also to break the structure of nutritionalAttributes', () => {
@@ -16,27 +15,6 @@ describe('nutritionalAttributes data files returns array', () => {
   })
 })
 
-const schema = {
-  properties: {
-    Nutrition: {
-      type: 'string'
-    },
-    Description: {
-      type: 'string'
-    },
-    ImpliedUnits: {
-      type: 'string'
-    }
-  },
-
-  required: ['Nutrition', 'Description', 'ImpliedUnits']
-}
-
-const example = {
-  Nutrition: 'K',
-  Description: 'Potassium, K',
-  ImpliedUnits: 'gram'
-}
 describe('testng for nutritionalAttribute json data schema', () => {
   it('validates nutritionalAttribute json schema', () => {
     expect(example).toMatchSchema(schema)
