+++ tests/pint.test.js
@@ -5,30 +5,14 @@ expect.extend(matchers)
 
 const staticData = require('../filesObjects')
 
+const schema = require('./examples/pint').schema
+const example = require('./examples/pint').example
+
 describe('this test prevents to any issues and problems, also to break the structure of pint data', () => {
   it('pint data files returns object', () => {
     expect(staticData.pint).not.toBe('')
   })
 })
-
-const schema = {
-  properties: {
-    name: { type: 'string' },
-    americanStandart: { type: 'string' },
-    americanStandartOunces: { type: 'string' },
-    volume: { type: 'string' },
-    weight: { type: 'string' }
-  },
-  required: [ 'name', 'americanStandart', 'americanStandartOunces', 'volume', 'weight' ]
-}
-
-const example = {
-  name: 'teaspoon',
-  americanStandart: '4 tablespoons',
-  americanStandartOunces: '2 ounces',
-  volume: '2 milliliters',
-  weight: '56.7 grams'
-}
 
 describe('test pint json schema', () => {
   it('validates pint json schema', () => {
