+++ tests/quart.test.js
@@ -5,29 +5,15 @@ expect.extend(matchers)
 
 const staticData = require('../filesObjects')
 
+const schema = require('./examples/quart').schema
+const example = require('./examples/quart').example
+
 describe('this test prevents to any issues and problems, also to break the structure of quart data', () => {
   it('quart data files returns object', () => {
     expect(staticData.quart).not.toBe('')
   })
 })
 
-const schema = {
-  properties: {
-    name: { type: 'string' },
-    americanStandart: { type: 'string' },
-    americanStandartOunces: { type: 'string' },
-    volume: { type: 'string' },
-    weight: { type: 'null' }
-  },
-  required: [ 'name', 'americanStandart', 'americanStandartOunces', 'volume', 'weight' ]
-}
-const example = {
-  name: 'teaspoon',
-  americanStandart: '4 cups,  2 pints',
-  americanStandartOunces: '32 ounces',
-  volume: '2 milliliters',
-  weight: null
-}
 describe('testng for quart json data schema', () => {
   it('validates quart json-schema', () => {
     expect(example).toMatchSchema(schema)
